Update findTemplate()
FindTemlate now requires a second optional parameter to supress the throwing of the exception

An exception has been thrown during the rendering of a template ("Error: In ZfcTwig\Twig\StackLoader::findTemplate(), you must accept a second argument that when set to "false" returns "false" instead of throwing an exception. Not supporting this argument is deprecated since version 1.27.